backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.0.0
razorpay==1.4.2
reportlab==4.0.4
pillow==10.0.1
python-decouple==3.8
psycopg2-binary==2.9.10
requests==2.31.0
dj-database-url==1.0.0
whitenoise==6.5.0
supabase==1.0.0
gunicorn==20.1.0